+++ backend/src/main/kotlin/com/otoki/internal/repository/AccountRepository.kt
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 /**
  * 거래처 마스터 Repository
  */
-interface AccountRepository : JpaRepository<Account, Long> {
+interface AccountRepository : JpaRepository<Account, Int> {
 
     /**
      * 거래처 외부키(SAP 코드)로 조회
@@ -16,7 +16,7 @@ interface AccountRepository : JpaRepository<Account, Long> {
     /**
      * 거래처 ID 목록으로 일괄 조회
      */
-    fun findByIdIn(ids: List<Long>): List<Account>
+    fun findByIdIn(ids: List<Int>): List<Account>
 
     /**
      * 거래처 sfid 목록으로 일괄 조회 (스케줄 → 거래처명 매핑용)
