+++ backend/src/main/kotlin/com/otoki/internal/repository/StaffReviewRepository.kt
@@ -3,4 +3,4 @@ package com.otoki.internal.repository
 import com.otoki.internal.entity.StaffReview
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface StaffReviewRepository : JpaRepository<StaffReview, Long>
+interface StaffReviewRepository : JpaRepository<StaffReview, Int>
