+++ backend/src/main/kotlin/com/otoki/internal/repository/UploadFileRepository.kt
@@ -3,4 +3,4 @@ package com.otoki.internal.repository
 import com.otoki.internal.entity.UploadFile
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface UploadFileRepository : JpaRepository<UploadFile, Long>
+interface UploadFileRepository : JpaRepository<UploadFile, Int>
